\name{annot}
\alias{annot}
\docType{data}
\title{Data file annot}
\description{Links probeset ID with gene_name}
\usage{data("annot")}
\format{
  A data frame with 5000 observations on the following 2 variables.
  \describe{
    \item{\code{probeset_id}}{a character vector}
    \item{\code{gene_name}}{a character vector}
  }
}
\source{
TTCA: An R Package for the identification of differentially expressed genes in time course microarray data
}
\examples{
data(annot)
}
\keyword{datasets}
