\name{annotation}
\alias{annotation}
\docType{data}
\title{Data file annotation}
\description{Gives the possibility to link the resulttable with further characteristics.}
\usage{data("annotation")}
\format{
  A data frame with 5000 observations on the following 2 variables.
  \describe{
    \item{\code{GeneOntology}}{a character vector}
    \item{\code{Property}}{a character vector}
  }
}
\source{
TTCA: An R Package for the identification of differentially expressed genes in time course microarray data
}
\examples{
data(annotation)
}
\keyword{datasets}
