\name{Control}
\alias{Control}
\docType{data}
\title{Control time course}
\description{
The second time course representing the control. Lung cancer cell line.}
\usage{data("Control")}
\format{
  A data frame with 5000 observations on the following 11 variables.
  \describe{
    \item{\code{B_C_2_0}}{a numeric vector}
    \item{\code{B_C_3_0}}{a numeric vector}
    \item{\code{B_C_2_0.5}}{a numeric vector}
    \item{\code{B_C_1_1}}{a numeric vector}
    \item{\code{B_C_3_4}}{a numeric vector}
    \item{\code{B_C_2_6}}{a numeric vector}
    \item{\code{B_C_1_24}}{a numeric vector}
    \item{\code{B_C_3_24}}{a numeric vector}
    \item{\code{B_C_1_48}}{a numeric vector}
    \item{\code{B_C_2_48}}{a numeric vector}
    \item{\code{B_C_3_48}}{a numeric vector}
  }
}
\source{
TTCA: An R Package for the identification of differentially expressed genes in time course microarray data
}
\examples{
data(Control)
}
\keyword{datasets}
