\name{EGF}
\alias{EGF}
\docType{data}
\title{EGF time course data}
\description{
The first time course representing the sample with epidermal growth factor stimulation. Lung cancer cell line.
}
\usage{data("EGF")}
\format{
  A data frame with 5000 observations on the following 16 variables.
  \describe{
    \item{\code{B_C_2_0}}{a numeric vector}
    \item{\code{B_C_3_0}}{a numeric vector}
    \item{\code{B_EGF_1_0.5}}{a numeric vector}
    \item{\code{B_EGF_3_0.5}}{a numeric vector}
    \item{\code{B_EGF_1_1}}{a numeric vector}
    \item{\code{B_EGF_3_2}}{a numeric vector}
    \item{\code{B_EGF_1_4}}{a numeric vector}
    \item{\code{B_EGF_1_6}}{a numeric vector}
    \item{\code{B_EGF_1_8}}{a numeric vector}
    \item{\code{B_EGF_2_12}}{a numeric vector}
    \item{\code{B_EGF_2_18}}{a numeric vector}
    \item{\code{B_EGF_1_24}}{a numeric vector}
    \item{\code{B_EGF_3_24}}{a numeric vector}
    \item{\code{B_EGF_1_48}}{a numeric vector}
    \item{\code{B_EGF_2_48}}{a numeric vector}
    \item{\code{B_EGF_3_48}}{a numeric vector}
  }
}
\source{
TTCA: An R Package for the identification of differentially expressed genes in time course microarray data
}
\examples{
data(EGF)
}
\keyword{datasets}
