% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_profile.R
\name{visualize_profile}
\alias{visualize_profile}
\title{Visualize Profile for a Given Gene}
\usage{
visualize_profile(
  gene_symbol,
  TTCA_result,
  grp1,
  grp1.time,
  grp2,
  grp2.time,
  annot = NA
)
}
\arguments{
\item{gene_symbol}{The symbol of the gene whose profile is to be visualized (character)}

\item{TTCA_result}{Result of TTCA analysis (data.frame)}

\item{grp1}{Data set with longitudinal sampled data (data.frame)}

\item{grp1.time}{Time points for data set 1 (vector like: c(0,0,0.5,1,2,4,6,8,12,12)}

\item{grp2}{Data set with longitudinal sampled data for comparison (data.frame)}

\item{grp2.time}{Time points for data set 2 (vector like: c(0,0,0.5,3,2,4,6,8,12,12,24)}

\item{annot}{Annotation for pictures and result (Data.frame with 2 columns with ID and GeneName). (Default: annot=NA)}
}
\value{
displays a plot of the gene expression profile for the gene. Returns the data frame TTCA p-values for the gene (if it exists, if not, returns
"Gene not analyzed by TTCA")
}
\description{
Visualize Profile for a Given Gene
}
